requirements.txt
uagents>=0.1.0  # Fetch.ai uAgents framework for Agentverse
requests>=2.31.0
boto3>=1.34.0
flask>=2.3.0
flask-cors>=4.0.0
geopy>=2.4.0
pyyaml>=6.0  # For agentverse_config.yaml parsing